Add generatorposition added indexing for generatorid on schema ast split out find_at_position some clean-up of publicity
