+++ psl/parser-database/src/names.rs
@@ -186,11 +186,11 @@ fn check_for_duplicate_properties<'a>(
 ) {
     tmp_names.clear();
     for arg in props {
-        if !tmp_names.insert(&arg.name.name) {
+        if !tmp_names.insert(&arg.name()) {
             ctx.push_error(DatamodelError::new_duplicate_config_key_error(
                 &format!("{} \"{}\"", top.get_type(), top.name()),
-                &arg.name.name,
-                arg.name.span,
+                arg.name(),
+                arg.identifier().span,
             ));
         }
     }
